Changed to lazy execution.  Moved PausableThreadPoolExecutor to its own class.
