+++ scattersphere-core/src/main/scala/com/scattersphere/core/util/execution/JobExecutor.scala
@@ -14,10 +14,6 @@
 package com.scattersphere.core.util.execution
 
 import java.util.concurrent.{CompletionException, _}
-import java.util.concurrent.BlockingQueue
-import java.util.concurrent.ThreadPoolExecutor
-import java.util.concurrent.TimeUnit
-import java.util.concurrent.locks.{Condition, ReentrantLock}
 import java.util.function.{Function => JavaFunction}
 
 import com.scattersphere.core.util._
@@ -37,7 +33,7 @@ import scala.collection.mutable
   */
 class JobExecutor(job: Job) {
 
-  private val executorService: PausableThreadPoolExecutor = new PausableThreadPoolExecutor()
+  private lazy val executorService: PausableThreadPoolExecutor = PausableThreadPoolExecutor()
 
   private val taskMap: mutable.HashMap[String, CompletableFuture[Void]] = new mutable.HashMap
   private val lockObject: Object = new Object
@@ -194,85 +190,3 @@ class InvalidTaskStateException(task: Task,
                                 status: TaskStatus,
                                 expected: TaskStatus)
   extends Exception(s"InvalidTaskStateException: task ${task.name} set to $status, expected $expected")
-
-/**
-  * A light wrapper around the {@link ThreadPoolExecutor}. It allows for you to pause execution and
-  * resume execution when ready. It is very handy for games that need to pause.
-  *
-  * (Please note, no license was specified when copied from GitHubGist, so this applies to the LICENSE-2.0
-  * as outlined in the start of this code.)
-  *
-  * @author Matthew A. Johnston (warmwaffles)
-  * @param corePoolSize    The size of the pool
-  * @param maximumPoolSize The maximum size of the pool
-  * @param keepAliveTime   The amount of time you wish to keep a single task alive
-  * @param unit            The unit of time that the keep alive time represents
-  * @param workQueue       The queue that holds your tasks
-  * @see { @link ThreadPoolExecutor#ThreadPoolExecutor(int, int, long, TimeUnit, BlockingQueue)}
-  */
-class PausableThreadPoolExecutor(val corePoolSize: Int = Runtime.getRuntime.availableProcessors(),
-                                 val maximumPoolSize: Int = Runtime.getRuntime.availableProcessors() * 10,
-                                 val keepAliveTime: Long = Long.MaxValue,
-                                 val unit: TimeUnit = TimeUnit.SECONDS,
-                                 val workQueue: BlockingQueue[Runnable] = new LinkedBlockingQueue[Runnable]())
-  extends ThreadPoolExecutor(corePoolSize, maximumPoolSize, keepAliveTime, unit, workQueue) {
-
-  private val lock: ReentrantLock = new ReentrantLock()
-  private val condition: Condition = lock.newCondition()
-  private var paused = false
-
-  /**
-    * @param thread   The thread being executed
-    * @param runnable The runnable task
-    * @see { @link ThreadPoolExecutor#beforeExecute(Thread, Runnable)}
-    */
-  override protected def beforeExecute(thread: Thread, runnable: Runnable): Unit = {
-    super.beforeExecute(thread, runnable)
-
-    lock.lock()
-
-    try {
-      while (paused) {
-        println("Awaiting lock release.")
-        condition.await
-        println("Lock released.")
-      }
-    } catch {
-      case _: InterruptedException => thread.interrupt()
-    } finally {
-      lock.unlock()
-    }
-  }
-
-  def isRunning: Boolean = !paused
-
-  def isPaused: Boolean = paused
-
-  /**
-    * Pause the execution
-    */
-  def pause(): Unit = {
-    println(s"PausableThreadPoolExecutor: Pausing.")
-    lock.lock()
-    try {
-      paused = true
-    } finally {
-      lock.unlock()
-    }
-  }
-
-  /**
-    * Resume pool execution
-    */
-  def resume(): Unit = {
-    println(s"PausableThreadPoolExecutor: Resuming.")
-    lock.lock()
-
-    try {
-      paused = false
-      condition.signalAll
-    } finally {
-      lock.unlock()
-    }
-  }
-}
